Introduce cross-building to newer Scala version (#4)
With the release of Scala 2.12, the projects I've been using the
circuit-breaker in are not able to update. Therefore I added a
cross-build setting to the sbt configuration for the current and
previous version of Scala. The tests still pass (after updating
scalatest).

It would be great to be able to use this library going forward with
an Scala update.